resolve conflicts

+++ SocialNetworkContentAnalyzerAPI/src/com/contentanalyzer/springservice/domain/Classification.java
@@ -51,7 +51,7 @@ public class Classification {
 			double clsLabel = smo.classifyInstance(testingData.instance(i));
 			System.out.println(testingData.instance(i));
 			System.out.println("clsLabel : " + clsLabel);
-			labeled.instance(i).setClassValue((int)clsLabel);
+			labeled.instance(i).setClassValue(clsLabel);
 			System.out.println(labeled.instance(i).classValue());
 			System.out.println("Class predicted: " + trainingData.classAttribute().value((int) clsLabel));	
 		}
